Move testutils to its own directory
Line up this package with all other packages, and ensure that the tests
don't try to import everything in `/pkg` (which will grow over time).

Tested: ran the e2e tests (skipped HNC_REPAIR tests since I have the
nomos-operator installed, and the demo failed for an unrelated reason,
but otherwise everything looks ok).
