+++ incubator/hnc/test/e2e/demo_test.go
@@ -2,7 +2,7 @@ package e2e
 
 import (
 	. "github.com/onsi/ginkgo"
-	. "sigs.k8s.io/multi-tenancy/incubator/hnc/pkg"
+	. "sigs.k8s.io/multi-tenancy/incubator/hnc/pkg/testutils"
 )
 
 var _ = Describe("Demo", func() {
