+++ incubator/hnc/test/e2e/issues_test.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	. "github.com/onsi/ginkgo"
-	. "sigs.k8s.io/multi-tenancy/incubator/hnc/pkg"
+	. "sigs.k8s.io/multi-tenancy/incubator/hnc/pkg/testutils"
 )
 
 var _ = Describe("Issues", func() {
